--- src/routes/guide/data.css
@@ -1,7 +1,0 @@
-.iconWrapper {
-    padding: 1rem 0;
-}
-
-:root {
-    --mrt-base-background-color: var(--mantine-color-green-0);
-}
